# backend/openapi.yaml
openapi: 3.1.0
info:
  title: UITGo Backend API
  version: 0.4.0
  description: |
    Minimal trip & auth API powering the UITGo rider/driver apps.
    Realtime updates stream over WebSocket at `ws://{host}/v1/trips/{tripId}/ws`.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/trips:
    get:
      summary: List trips for current user
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [rider, driver]
            default: rider
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: page
          in: query
          required: false
          description: Optional 1-based page index (takes precedence over limit/offset)
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: false
          description: Page size when using the page parameter
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Trip page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripListResponse'
        '401':
          description: Unauthorized
    post:
      summary: Create a trip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
      responses:
        '201':
          description: Trip created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '401':
          description: Unauthorized
  /v1/trips/{id}:
    get:
      summary: Get trip details
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Trip fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '404':
          description: Trip not found
    patch:
      summary: Update trip status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Trip updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '400':
          description: Invalid request
        '404':
          description: Trip not found
  /v1/trips/{id}/ws:
    get:
      summary: Trip realtime channel
      description: |
        Upgrades to WebSocket. Drivers send `{"type":"location","lat":0,"lng":0}` messages.
        Riders receive broadcasts including driver locations and status changes.
        Optional query parameters `role` and `userId` can be supplied when custom headers
        are not available (e.g. browser clients).
      parameters:
        - $ref: '#/components/parameters/TripId'
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [rider, driver]
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        '101':
          description: WebSocket Upgrade
...
  /wallet:
    get:
      summary: '[Deprecated] Get rider wallet summary'
      description: This legacy endpoint is kept for backward compatibility. Use `/v1/wallet` instead.
      deprecated: true
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Wallet summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '401':
          description: Unauthorized
  /v1/wallet:
    get:
      summary: Get wallet balance and rewards
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Wallet summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '401':
          description: Unauthorized
  /v1/wallet/transactions:
    get:
      summary: List wallet transactions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletTransactionListResponse'
        '401':
          description: Unauthorized
  /v1/wallet/topup:
    post:
      summary: Mock wallet top-up
      description: Validates the amount and instantly credits the rider wallet.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletTopUpRequest'
      responses:
        '201':
          description: Updated wallet summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          description: Invalid amount
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TripId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    TripResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        riderId:
          type: string
        driverId:
          type: string
          nullable: true
        serviceId:
          type: string
        originText:
          type: string
        destText:
          type: string
        status:
          $ref: '#/components/schemas/TripStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLocation:
          $ref: '#/components/schemas/LocationUpdate'
          nullable: true
    TripListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TripResponse'
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    CreateTripRequest:
      type: object
      required:
        - originText
        - destText
        - serviceId
      properties:
        originText:
          type: string
        destText:
          type: string
        serviceId:
          type: string
    UpdateStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/TripStatus'
    TripStatus:
      type: string
      enum: [requested, accepted, arriving, in_ride, completed, cancelled]
    LocationUpdate:
      type: object
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
...
    Wallet:
      type: object
      properties:
        balance:
          type: integer
          format: int64
        rewardPoints:
          type: integer
        updatedAt:
          type: string
          format: date-time
    WalletTopUpRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: integer
          format: int64
          minimum: 10000
          description: Amount in VND to credit immediately.
    WalletTransaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [topup, reward, deduction]
        amount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    WalletTransactionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WalletTransaction'
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
...